+++ src/typeahead.jsx
@@ -45,6 +45,7 @@ export default class Typeahead extends Component {
 
     this.state = {
       focused: null,
+      hovered: null,
       menuOpen: false,
       options: props.defaultValue ? [props.defaultValue] : [],
       query: props.defaultValue,
@@ -62,7 +63,7 @@ export default class Typeahead extends Component {
     this.handleOptionFocusOut = this.handleOptionFocusOut.bind(this)
     this.handleOptionFocus = this.handleOptionFocus.bind(this)
     this.handleOptionMouseDown = this.handleOptionMouseDown.bind(this)
-    this.handleOptionMouseMove = this.handleOptionMouseMove.bind(this)
+    this.handleOptionMouseEnter = this.handleOptionMouseEnter.bind(this)
     this.handleOptionMouseOut = this.handleOptionMouseOut.bind(this)
 
     this.handleInputBlur = this.handleInputBlur.bind(this)
@@ -216,27 +217,21 @@ export default class Typeahead extends Component {
   handleOptionFocus (idx) {
     this.setState({
       focused: idx,
+      hovered: null,
       selected: idx
     })
   }
 
-  handleOptionMouseMove (idx) {
+  handleOptionMouseEnter (evt, idx) {
     this.setState({
-      focused: idx
+      hovered: idx
     })
   }
 
   handleOptionMouseOut (evt, idx) {
-    const previousOption = this.elementRefs[idx - 1]
-    const nextOption = this.elementRefs[idx + 1]
-    const toElement = evt.toElement
-    const movingToAnotherOption = toElement === previousOption || toElement === nextOption
-    const focusBackOnSelectedOption = !movingToAnotherOption
-    if (focusBackOnSelectedOption) {
-      this.setState({
-        focused: this.state.selected
-      })
-    }
+    this.setState({
+      hovered: null
+    })
   }
 
   handleOptionClick (evt, idx) {
@@ -325,7 +320,7 @@ export default class Typeahead extends Component {
 
   render () {
     const { cssNamespace, displayMenu, id, minLength, name } = this.props
-    const { focused, menuOpen, options, query, selected } = this.state
+    const { focused, hovered, menuOpen, options, query, selected } = this.state
     const autoselect = this.hasAutoselect()
 
     const inputFocused = focused === -1
@@ -403,7 +398,7 @@ export default class Typeahead extends Component {
     const Option = ({ dangerouslySetInnerHTML, idx }) => {
       const cn = `${cssNamespace}__option`
       const showFocused = focused === -1 ? selected === idx : focused === idx
-      const cnModFocused = showFocused ? ` ${cn}--focused` : ''
+      const cnModFocused = showFocused && hovered === null ? ` ${cn}--focused` : ''
       const cnModOdd = (idx % 2) ? ` ${cn}--odd` : ''
       const cns = `${cn}${cnModFocused}${cnModOdd}`
       return <li
@@ -414,7 +409,7 @@ export default class Typeahead extends Component {
         onClick={(evt) => this.handleOptionClick(evt, idx)}
         onFocusOut={(evt) => this.handleOptionFocusOut(evt, idx)}
         onMouseDown={this.handleOptionMouseDown}
-        onMouseMove={() => this.handleOptionMouseMove(idx)}
+        onMouseEnter={(evt) => this.handleOptionMouseEnter(evt, idx)}
         onMouseOut={(evt) => this.handleOptionMouseOut(evt, idx)}
         role='option'
         tabindex='-1'